Refactor attachment rendering in DetailsTab components
- Removed the `useGroupedAttachments` hook and replaced it with inline logic to display attachments directly in Files, Links, Media, and Voices components.
- Updated the rendering logic to conditionally show month headers based on attachment dates, improving clarity and organization.
- Cleaned up import statements and ensured consistent usage of the `MonthHeader` component across all relevant sections.
