+++ src/components/ChannelDetails/DetailsTab/shared/index.ts
@@ -1,3 +1,2 @@
 export { default as MonthHeader } from './MonthHeader'
-export { useGroupedAttachments } from './useGroupedAttachments'
 export type { GroupedAttachment } from './types'
